@@ -5,43 +5,32 @@
   "dependencies": {
     "@ant-design/icons": "^5.0.1",
     "@reduxjs/toolkit": "^1.9.1",
-    "@web3modal/ethereum": "^2.1.1",
-    "@web3modal/react": "^2.1.1",
-    "antd": "^5.1.7",
-    "axios": "^0.21.4",
+    "@vitejs/plugin-react": "^4.2.0",
+    "@web3modal/wagmi": "^3.3.2",
+    "antd": "^5.11.4",
+    "axios": "^1.6.2",
     "base-64": "^1.0.0",
     "email-validator": "^2.0.4",
-    "env-cmd": "^10.1.0",
-    "ethers": "^5.7.2",
-    "i18next": "^19.9.2",
-    "i18next-browser-languagedetector": "^6.1.8",
-    "i18next-xhr-backend": "^3.2.2",
     "lodash": "^4.17.21",
-    "onfido-sdk-ui": "^10.0.2",
     "react": "^18.2.0",
     "react-awesome-reveal": "^3.8.1",
     "react-confetti": "^6.1.0",
     "react-confetti-explosion": "^2.1.2",
     "react-cookie-consent": "^8.0.1",
     "react-dom": "^18.2.0",
-    "react-i18next": "^11.18.6",
-    "react-markdown": "^8.0.7",
-    "react-player": "^2.13.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.6.1",
-    "react-scripts": "5.0.1",
-    "rehype-raw": "^6.1.1",
-    "styled-components": "^5.3.6",
-    "typescript": "^4.9.3",
-    "wagmi": "^0.12.12",
-    "web3": "^1.9.0",
+    "styled-components": "^6.1.1",
+    "typescript": "^5.3.2",
+    "vite": "^5.0.2",
+    "vite-tsconfig-paths": "^4.2.1",
+    "wagmi": "^1.4.7",
+    "web3": "^4.2.2",
     "zod": "^3.21.4"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -62,9 +51,9 @@
     "@trivago/prettier-plugin-sort-imports": "^4.1.1",
     "@types/base-64": "^1.0.0",
     "@types/lodash": "^4.14.191",
-    "@types/node": "^15.14.9",
+    "@types/node": "^20.10.0",
     "@types/react": "^17.0.52",
-    "@types/react-dom": "^17.0.18",
+    "@types/react-dom": "^18.2.17",
     "@types/react-router-dom": "^5.3.3",
     "@types/styled-components": "^5.1.26",
     "prettier": "2.8.7"
