@@ -16,32 +16,32 @@ export const web3Factory = (
 ): { web3: Web3; soulboundContract: string } => {
   const supportedChains: Chains = {
     [SupportedChainId.POLYGON]: {
-      provider: ENV.REACT_APP_POLYGON_ALCHEMY_URL,
-      soulboundContract: ENV.REACT_APP_POLYGON_SOULBOUND_CONTRACT,
+      provider: ENV.VITE_APP_POLYGON_ALCHEMY_URL,
+      soulboundContract: ENV.VITE_APP_POLYGON_SOULBOUND_CONTRACT,
     },
     [SupportedChainId.SEPOLIA]: {
-      provider: ENV.REACT_APP_SEPOLIA_ALCHEMY_URL,
-      soulboundContract: ENV.REACT_APP_SEPOLIA_SOULBOUND_CONTRACT,
+      provider: ENV.VITE_APP_SEPOLIA_ALCHEMY_URL,
+      soulboundContract: ENV.VITE_APP_SEPOLIA_SOULBOUND_CONTRACT,
     },
     [SupportedChainId.POLYGON_MUMBAI]: {
-      provider: ENV.REACT_APP_MUMBAI_INFURA_URL,
-      soulboundContract: ENV.REACT_APP_MUMBAI_SOULBOUND_CONTRACT,
+      provider: ENV.VITE_APP_MUMBAI_INFURA_URL,
+      soulboundContract: ENV.VITE_APP_MUMBAI_SOULBOUND_CONTRACT,
     },
     // [SupportedChainId.SCROLL_ALPHA]: {
     //   provider: "https://alpha-rpc.scroll.io/l2",
-    //   soulboundContract: ENV.REACT_APP_SCROLL_SOULBOUND_CONTRACT,
+    //   soulboundContract: ENV.VITE_APP_SCROLL_SOULBOUND_CONTRACT,
     // },
     [SupportedChainId.SCROLL_SEPOLIA]: {
       provider: "https://sepolia-rpc.scroll.io",
-      soulboundContract: ENV.REACT_APP_SCROLL_SEPOLIA_SOULBOUND_CONTRACT,
+      soulboundContract: ENV.VITE_APP_SCROLL_SEPOLIA_SOULBOUND_CONTRACT,
     },
     [SupportedChainId.BNB]: {
       provider: "https://bsc-dataseed.bnbchain.org",
-      soulboundContract: ENV.REACT_APP_BNB_SOULBOUND_CONTRACT,
+      soulboundContract: ENV.VITE_APP_BNB_SOULBOUND_CONTRACT,
     },
     [SupportedChainId.BNB_TESTNET]: {
       provider: "https://bsc-testnet-dataseed.bnbchain.org",
-      soulboundContract: ENV.REACT_APP_BNB_TESTNET_SOULBOUND_CONTRACT,
+      soulboundContract: ENV.VITE_APP_BNB_TESTNET_SOULBOUND_CONTRACT,
     },
   };
 
