@@ -16,7 +16,7 @@ export const BotInvitationSection = () => {
         <RedirectDiscordLink
           target="_blank"
           rel="noopener noreferrer"
-          href={ENV.REACT_APP_DISCORD_BOT_INVITATION_URL}
+          href={ENV.VITE_APP_DISCORD_BOT_INVITATION_URL}
         >
           Invite Sunscreen Bot
         </RedirectDiscordLink>
