@@ -28,7 +28,7 @@ const SuncreenVerification = ({ userData }: Props) => {
   const dispatch = useAppDispatch();
 
   const handleDiscordConnect = () => {
-    window.location.href = ENV.REACT_APP_DISCORD_OAUTH_URL;
+    window.location.href = ENV.VITE_APP_DISCORD_OAUTH_URL;
   };
 
   const onfidoRedirect = async () => {
