@@ -1,3 +1,3 @@
 import { ENV } from "env";
 
-export const API_URL = ENV.REACT_APP_BASE_URL || "http://localhost:3001/";
+export const API_URL = ENV.VITE_APP_BASE_URL || "http://localhost:3001/";
